Move .h to .hpp
