+++ MeshFeatureTracking/src/UI/MeshFeatureTrackingUI.cpp
@@ -1,4 +1,4 @@
-#include "MeshFeatureTrackingUI.h"
+#include "MeshFeatureTrackingUI.hpp"
 
 #include <iostream>
